Fix actoris-sidecar compilation for Windows
- Add windows-sys dependency for Windows process stats
- Fix MeteringCollector::new() to accept Duration parameter
- Change prometheus register() return type to avoid ActorisError conversion
- Fix metering session method calls (start_session, complete)
- Fix memory_bytes type cast (usize to u64)
- Add Validation variant to ActorisError
- Partial fixes to actoris-onebill (type alignments)

+++ crates/actoris-sidecar/src/metering/mod.rs
@@ -422,7 +422,7 @@ fn get_process_stats() -> Result<(u64, u64), MeteringError> {
             0
         };
 
-        Ok((cpu_time_us, memory_bytes))
+        Ok((cpu_time_us, memory_bytes as u64))
     }
 }
